trade-copilot/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
pydantic-settings>=2.2
python-dotenv>=1.0
httpx>=0.27
anthropic>=0.40
openai>=1.30
chromadb>=0.4.24
sentence-transformers>=2.7
rank-bm25>=0.2.2
pypdf>=4.0
python-docx>=1.1
beautifulsoup4>=4.12
markdownify>=0.11
tiktoken>=0.7
langdetect>=1.0
PyYAML>=6.0
rich>=13.0
click>=8.1
